% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{methods}
\name{dbColumnInfo,Data.WorldResult-method}
\alias{dbColumnInfo,Data.WorldResult-method}
\title{Get the column types of a query result.}
\usage{
\S4method{dbColumnInfo}{Data.WorldResult}(res, ...)
}
\arguments{
\item{res}{A \code{Data.WorldResult} object. A data
frame with the query results can be obtained by
calling \code{\link{dbFetch}} on the result.}
}
\value{
A data frame with columns "name" and
"type" that list the result's columns and their
types.
}
\description{
Get the column types of a query result.
}
