% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{methods}
\name{dbDisconnect,Data.WorldConnection-method}
\alias{dbDisconnect,Data.WorldConnection-method}
\title{A disconnect method for Data.World connections.}
\usage{
\S4method{dbDisconnect}{Data.WorldConnection}(conn, ...)
}
\arguments{
\item{conn}{A \code{Data.WorldConnection} object, as created
by \code{\link[dwDBI]{dbConnect}}.}
}
\description{
This is a placeholder that doesn't do anything.
}
