% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\docType{methods}
\name{dbExistsTable,Data.WorldConnection,character-method}
\alias{dbExistsTable,Data.WorldConnection,character-method}
\title{Check if a table exists in a data.world dataset.}
\usage{
\S4method{dbExistsTable}{Data.WorldConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A \code{Data.WorldConnection} object, as created
by \code{\link[dwDBI]{dbConnect}}.}

\item{name}{The name of the table to check. The comparison
is case-insensitive.}
}
\value{
TRUE if the table is in the dataset, FALSE if not.
}
\description{
Check if a table exists in a data.world dataset.
}
\seealso{
\code{\link[dwapi]{list_tables}}

#' @export
}
