% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{methods}
\name{dbFetch,Data.WorldResult-method}
\alias{dbFetch,Data.WorldResult-method}
\title{Get a data frame with results of a sent query.}
\usage{
\S4method{dbFetch}{Data.WorldResult}(res, n = NULL, ...)
}
\arguments{
\item{res}{a \code{Data.WorldResult} object, resulting
from a call to \code{\link{dbSendQuery}}.}

\item{n}{Not used. All rows are fetched.}
}
\value{
A data frame.
}
\description{
Get a data frame with results of a sent query.
}
\examples{
\dontrun{
conn <- dbConnect(Data.World(), 'johndoe/petstore')
old_cats_res <- dbSendQuery(conn,
                            'SELECT name, age, color
                             FROM pets
                             WHERE species=='cat' AND age > 10')
old_cats <- dbFetch(old_cats_res)
}

}
