% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{methods}
\name{dbGetInfo,Data.WorldConnection-method}
\alias{dbGetInfo,Data.WorldConnection-method}
\title{Get information about a data.world dataset.}
\usage{
\S4method{dbGetInfo}{Data.WorldConnection}(dbObj, ...)
}
\arguments{
\item{conn}{A \code{Data.WorldConnection} object, as created
by \code{\link[dwDBI]{dbConnect}}.}
}
\value{
A named list with the following entries:
\itemize{
  \item owner
  \item id
  \item title
  \item visibility
  \item updated
  \item created
  \item status
  \item description
  \item summary
  \item tags
  \item license
  \item files
  }
}
\description{
Get information about a data.world dataset.
}
