% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\docType{methods}
\name{dbListFields,Data.WorldConnection,character-method}
\alias{dbListFields,Data.WorldConnection,character-method}
\title{List the columns of a table in a data.world dataset.}
\usage{
\S4method{dbListFields}{Data.WorldConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A \code{Data.WorldConnection} object, as created
by \code{\link[dwDBI]{dbConnect}}.}

\item{name}{The name of the table to check. The comparison
is case-insensitive.}
}
\value{
A character vector of field names.
}
\description{
List the columns of a table in a data.world dataset.
}
\seealso{
\code{\link[dwapi]{get_table_schema}}
}
