% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\docType{methods}
\name{dbListTables,Data.WorldConnection-method}
\alias{dbListTables,Data.WorldConnection-method}
\title{List the tables in the data.world dataset connection.}
\usage{
\S4method{dbListTables}{Data.WorldConnection}(conn, ...)
}
\arguments{
\item{conn}{A \code{Data.WorldConnection} object, as created
by \code{\link[dwDBI]{dbConnect}}.}
}
\value{
A character vector of table names.
}
\description{
List the tables in the data.world dataset connection.
}
\seealso{
\code{\link[dwapi]{list_tables}}.
}
