% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\docType{methods}
\name{dbReadTable,Data.WorldConnection,character-method}
\alias{dbReadTable,Data.WorldConnection,character-method}
\title{Read an entire data.world table into a data frame.}
\usage{
\S4method{dbReadTable}{Data.WorldConnection,character}(conn, name)
}
\arguments{
\item{conn}{A \code{Data.WorldConnection} object, as created
by \code{\link[dwDBI]{dbConnect}}.}

\item{name}{Name of the table to read.}
}
\value{
A data frame.
}
\description{
Read an entire data.world table into a data frame.
}
\examples{
\dontrun{
conn <- dbConnect(Data.World(), dataset='johndoe/petstore')
customers <- dbReadTable(conn, customers)
}

}
\seealso{
\code{\link[dwapi]{download_table_as_data_frame}}.
}
