% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\docType{methods}
\name{dbSendQuery,Data.WorldConnection-method}
\alias{dbSendQuery,Data.WorldConnection-method}
\title{Send a query to a data.world dataset.}
\usage{
\S4method{dbSendQuery}{Data.WorldConnection}(conn, statement, ...)
}
\arguments{
\item{conn}{A \code{Data.WorldConnection} object, as created
by \code{\link[dwDBI]{dbConnect}}.}

\item{statement}{A SQL query string. The query
can be parameterized.}

\item{params}{Any parameters to pass into the
query before running.}
}
\value{
A \code{Data.WorldResult} object. A data
frame with the query results can be obtained by
calling \code{\link{dbFetch}} on the result.
}
\description{
Send a query to a data.world dataset.
}
\seealso{
\code{\link[dwapi]{sql}}.
}
