% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\docType{methods}
\name{dbWriteTable,Data.WorldConnection,character,character-method}
\alias{dbWriteTable,Data.WorldConnection,character,character-method}
\title{Upload a local file to a data.world dataset.}
\usage{
\S4method{dbWriteTable}{Data.WorldConnection,character,character}(conn,
  name = NULL, value, ...)
}
\arguments{
\item{conn}{A \code{Data.WorldConnection} object, as created
by \code{\link[dwDBI]{dbConnect}}.}

\item{name}{What the uploaded file should be named. If the
name does not end in '.csv', it will be added.}

\item{value}{A path to local file.}
}
\value{
TRUE if the file was written successfully.
}
\description{
Upload a local file to a data.world dataset.
}
\seealso{
\code{\link[dwapi]{upload_file}}
}
