% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\docType{methods}
\name{dbWriteTable,Data.WorldConnection,character,data.frame-method}
\alias{dbWriteTable,Data.WorldConnection,character,data.frame-method}
\title{Upload a data frame as a CSV file a data.world dataset.}
\usage{
\S4method{dbWriteTable}{Data.WorldConnection,character,data.frame}(conn, name,
  value, ...)
}
\arguments{
\item{conn}{A \code{Data.WorldConnection} object, as created
by \code{\link[dwDBI]{dbConnect}}.}

\item{name}{What the uploaded file should be named. If the
name does not end in '.csv', it will be added.}

\item{value}{A data frame to upload to the dataset.}
}
\value{
TRUE if the file was written successfully.
}
\description{
Upload a data frame as a CSV file a data.world dataset.
}
