% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{dw_connect}
\alias{dw_connect}
\title{Connect to a dataset like it's a database.}
\usage{
dw_connect(dataset)
}
\arguments{
\item{dataset}{The name of a data.world dataset.}
}
\value{
A "handle" in the form of Data.WorldConnection object.
}
\description{
Connect to a dataset like it's a database.
}
\note{
Even though it doesn't make an API call, this function
will throw an error if your data.world API token has not been
configured. See \code{\link[dwapi]{configure}}.
This ensures that any subsequent queries sent over
this connection will work.
}
\examples{
\dontrun{
conn <- dw_connect('johndoe/petstore')
dbGetQuery(conn, "SELECT * FROM pets WHERE species = 'dog'")
}

}
