% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{dw_dataset_viewer}
\alias{dw_dataset_viewer}
\title{View a data.world dataset in the RStudio Connections panel.}
\usage{
dw_dataset_viewer(dataset)
}
\arguments{
\item{dataset}{The name of a data.world dataset.}
}
\description{
View a data.world dataset in the RStudio Connections panel.
}
